csvqb/pyproject.toml
[tool.poetry]
name = "csvqb"
version = "0.1.0"
description = "A tool to generate RDF Data Cube style CSV-W cubes from tidy CSV files. Part of the csvcubed family."
authors = ["Integrated Data Service - Dissemination <example@example.com>"]
license = "Apache-2.0"
readme="csvqb/README.md"
exclude = ["docs/*", "tests/*"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "~=1.3.3"
rdflib-jsonld = "==0.6.1"
pydantic = {git = "https://github.com/robons/pydantic.git", develop = true}

[tool.poetry.dev-dependencies]
devtools = {path = "./../devtools", develop = true}
# The sharedmodels library is currently listed as a dev dependency so that we can reference it by its relative
# path and continue to edit code side-by-side
# The sharedmodels dependency is to become a regular dependency when the packages are split out into separate
# repositories in Issue #205
sharedmodels = {path = "./../sharedmodels", develop = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
pythonVersion="3.9"
exclude = [
    ".venv",
    "**/tests"
]